pyproject.toml
[project]
name = "ha-xenopixel"
version = "0.1.0"
description = "Home Assistant integration for Xenopixel V3 lightsabers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13.2"
authors = [
    {name = "Ollec"}
]
keywords = ["home-assistant", "bluetooth", "ble", "lightsaber", "xenopixel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

# Runtime dependencies (pinned versions, also listed in manifest.json)
dependencies = [
    "bleak==2.1.1",
    "bleak-retry-connector==4.4.4",
]

# UV dependency groups (use: uv sync --group dev)
# Note: pytest-homeassistant-custom-component pins homeassistant version, so don't pin HA separately
[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.14.0",
    "mypy>=1.15.0",
    "pre-commit>=4.0.0",
    "pytest-homeassistant-custom-component>=0.13.300",
]

[project.urls]
Homepage = "https://github.com/Ollec/ha-xenopixel"
Repository = "https://github.com/Ollec/ha-xenopixel"
Issues = "https://github.com/Ollec/ha-xenopixel/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["custom_components*"]

# Ruff configuration
[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001", "S101"]
"tools/*" = ["ARG001", "ARG005", "E722", "F841", "SIM102", "SIM105"]

[tool.ruff.lint.isort]
known-first-party = ["custom_components.xenopixel"]

# Ruff formatter (replaces Black)
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy configuration
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "bleak.*",
    "bleak_retry_connector.*",
    "homeassistant.*",
    "pytest_homeassistant_custom_component.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["custom_components/xenopixel"]
branch = true
omit = [
    "custom_components/xenopixel/config_flow.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 80
